refactor(analysis): simplify performance validation
